moved hist function into its own .py file
